style: update call to action component
