--- src/components/CallToAction.md
@@ -1,8 +1,0 @@
-```js padded
-    <CallToAction 
-        heading="Find out how Law Centres can help"
-        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
-        buttonText="Get Help"
-        link="/get-hep"
-    />
-```
