+++ src/components/CallToAction/CallToAction.mdx
@@ -1,0 +1,17 @@
+import { Canvas, Meta, Story } from '@storybook/blocks'
+
+import * as CallToActionStories from './CallToAction.stories'
+
+<Meta of={CallToActionStories} />
+
+# Call To Action
+
+The Call To Action component is used when you want to bring the user's focused attention to brief, but important, messages and encourage the user to take a specific action.
+
+## Overview
+
+This button is our default Call To Action.
+
+<Canvas>
+  <Story of={CallToActionStories.Default} />
+</Canvas>
